+++ tests/test_shipment.py
@@ -188,9 +188,9 @@ def test_shipment_lowest_smartrate(basic_shipment):
     shipment = easypost.Shipment.create(**basic_shipment)
 
     # Test lowest smartrate with valid filters
-    lowest_smartrate_filters = shipment.lowest_smartrate(delivery_days=1, delivery_accuracy="percentile_90")
-    assert lowest_smartrate_filters["service"] == "Priority"
-    assert lowest_smartrate_filters["rate"] == 7.37
+    lowest_smartrate_filters = shipment.lowest_smartrate(delivery_days=2, delivery_accuracy="percentile_90")
+    assert lowest_smartrate_filters["service"] == "First"
+    assert lowest_smartrate_filters["rate"] == 5.49
     assert lowest_smartrate_filters["carrier"] == "USPS"
 
     # Test lowest smartrate with invalid filters (should error due to strict delivery_days)
@@ -211,10 +211,10 @@ def test_shipment_get_lowest_smartrate(basic_shipment):
 
     # Test lowest smartrate with valid filters
     lowest_smartrate_filters = easypost.Shipment.get_lowest_smartrate(
-        smartrates, delivery_days=1, delivery_accuracy="percentile_90"
+        smartrates, delivery_days=2, delivery_accuracy="percentile_90"
     )
-    assert lowest_smartrate_filters["service"] == "Priority"
-    assert lowest_smartrate_filters["rate"] == 7.37
+    assert lowest_smartrate_filters["service"] == "First"
+    assert lowest_smartrate_filters["rate"] == 5.49
     assert lowest_smartrate_filters["carrier"] == "USPS"
 
     # Test lowest smartrate with invalid filters (should error due to strict delivery_days)
@@ -244,3 +244,38 @@ def test_generate_form(one_call_buy_shipment, rma_form_options):
 
     assert form.form_type == form_type
     assert form.form_url is not None
+
+
+@pytest.mark.vcr()
+def test_shipment_create_with_carbon_offset(carbon_offset_shipment):
+    shipment = easypost.Shipment.create(with_carbon_offset=True, **carbon_offset_shipment)
+
+    assert isinstance(shipment, easypost.Shipment)
+    assert shipment.rates is not None
+    assert all(rate.carbon_offset is not None for rate in shipment.rates)
+
+
+@pytest.mark.vcr()
+def test_shipment_buy_with_carbon_offset(carbon_offset_shipment):
+    shipment = easypost.Shipment.create(**carbon_offset_shipment)
+    shipment.buy(rate=shipment.lowest_rate(), with_carbon_offset=True)
+
+    assert isinstance(shipment, easypost.Shipment)
+    assert any(fee.type == "CarbonOffsetFee" for fee in shipment.fees)
+
+
+@pytest.mark.vcr()
+def test_shipment_one_call_buy_with_carbon_offset(carbon_offset_shipment_one_call_buy):
+    shipment = easypost.Shipment.create(with_carbon_offset=True, **carbon_offset_shipment_one_call_buy)
+
+    assert isinstance(shipment, easypost.Shipment)
+    assert all(rate.carbon_offset is not None for rate in shipment.rates)
+
+
+@pytest.mark.vcr()
+def test_shipment_rerate_with_carbon_offset(carbon_offset_shipment_one_call_buy):
+    shipment = easypost.Shipment.create(**carbon_offset_shipment_one_call_buy)
+
+    new_carbon_rates = shipment.regenerate_rates(with_carbon_offset=True)
+
+    assert all(rate.carbon_offset is not None for rate in new_carbon_rates.rates)